move LoopStatement.java to entities module + add type
