+++ score-lang-compiler/src/main/java/org/openscore/lang/compiler/SlangTextualKeys.java
@@ -35,7 +35,7 @@ public interface SlangTextualKeys {
 
     //task
     String DO_KEY = "do";
-    String LOOP_KEY = "loop";
+    String FOR_KEY = "for";
     String PUBLISH_KEY = "publish";
     String NAVIGATION_KEY = "navigate";
