+++ score-lang-compiler/src/main/java/org/openscore/lang/compiler/scorecompiler/ExecutionStepFactory.java
@@ -46,7 +46,7 @@ public class ExecutionStepFactory {
         Validate.notNull(preTaskData, "preTaskData is null");
         Map<String, Serializable> actionData = new HashMap<>();
         actionData.put(ScoreLangConstants.TASK_INPUTS_KEY, preTaskData.get(SlangTextualKeys.DO_KEY));
-        actionData.put(ScoreLangConstants.FOR_KEY, preTaskData.get(ScoreLangConstants.FOR_KEY));
+        actionData.put(ScoreLangConstants.LOOP_KEY, preTaskData.get(SlangTextualKeys.FOR_KEY));
         actionData.put(ScoreLangConstants.HOOKS, "TBD"); //todo add implementation for user custom hooks
         actionData.put(ScoreLangConstants.NODE_NAME_KEY, taskName);
         actionData.put(ScoreLangConstants.REF_ID, refId);
